Fix rounding error on total value calculation

+++ public/viewjs/stockoverview.js
@@ -229,7 +229,7 @@ function RefreshStatistics()
 				var valueSum = 0;
 				result.forEach(element =>
 				{
-					valueSum += parseInt(element.value);
+					valueSum += parseFloat(element.value);
 				});
 				$("#info-current-stock").text(__n(result.length, '%s Product', '%s Products') + ", " + __t('%s total value', valueSum.toLocaleString(undefined, { style: "currency", currency: Grocy.Currency })));
 			}
